Updated unsupported method test
