+++ test/404s.test.js
@@ -927,38 +927,6 @@ test('log debug for 404', t => {
   })
 })
 
-test('Unsupported method', t => {
-  t.plan(5)
-
-  const fastify = Fastify()
-
-  fastify.get('/', function (req, reply) {
-    reply.send({ hello: 'world' })
-  })
-
-  t.tearDown(fastify.close.bind(fastify))
-
-  fastify.listen(0, err => {
-    t.error(err)
-
-    fastify.inject({
-      method: 'PROPFIND',
-      url: '/'
-    }, (err, res) => {
-      t.error(err)
-      t.strictEqual(res.statusCode, 404)
-
-      sget({
-        method: 'PROPFIND',
-        url: 'http://localhost:' + fastify.server.address().port
-      }, (err, response, body) => {
-        t.error(err)
-        t.strictEqual(response.statusCode, 404)
-      })
-    })
-  })
-})
-
 test('recognizes errors from the http-errors module', t => {
   t.plan(5)
